First attempt ready for use
ThingControllerBase is now a library with wired and wireless modes built in and controlled by defines
Code is ready for initial use
