+++ .gitignore
@@ -24,3 +24,4 @@ fp-info-cache
 # Exported BOM files
 *.xml
 *.csv
+NewDoorController/data/Network.json
